Extract notification-related code from MessagingController to NotificationController
